pyproject.toml
[project]
name = "full-mcp-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.128.0",
    "langchain>=1.2.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-openai>=1.1.6",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.2",
]
